Add UI Resources prefabs and update UI scene
This commit introduces two new UI Resources prefabs: ResourceGrid and ResourcePanel along with their associated meta files. Additions also include the creation of a Canvas GameObject in the UI scene. A prefab instance was added to the existing Canvas GameObject and the UI scene was updated with additional changes to the RectTransform.