Add "hasReachedTarget" property to Unit class
The new property "hasReachedTarget" has been added to the Unit class to better track the status of a unit's movement or interaction with a target. The property is updated in UnitAI based on the proximity condition and is used in "MyCursor.cs" for different types of units. This change improves code clarity and allows for future features that depend on a unit's state with respect to the target.